ASoC: msm-lsm-client: Use freezable wait for EVENT_STATUS ioctl
The EVENT_STATUS ioctl waits until the underlying hardware raises an
interrupt. It is required for this ioctl to wait even in suspend state.
Using wait_event_interruptible causes issue as the userspace may try to
call the ioctl again. As this wait is needed even during suspended
state, fix the interruptible issues by using wait_event_freezable
so that the system can be frozen.

CRs-fixed: 734125
Change-Id: I1515b110ef0b39a6c021b1921179b61b4bc4909c
Signed-off-by: Bhalchandra Gajare <gajare@codeaurora.org>